+++ src/components/tree/OpenedIcon.tsx
@@ -33,6 +33,7 @@ const OpenedIcon: FC<Props> = ({ panelsNumbers, nodeType }) => {
     <Tooltip>
       <TooltipTrigger asChild>
         <div
+          data-cy="tree-node-marker"
           className="bg-primary w-2 h-2 hover:scale-130 transition-all duration-200 rounded-full"
           onClick={(e) => scrollToPanel(e)}
         ></div>