+++ src/utils/config/config.ts
@@ -122,7 +122,7 @@ function validateTheme(input: any): ValidationResult<TidoConfig['theme']> {
       typeof input === 'object' && typeof input['primaryColor'] === 'string'
         ? input['primaryColor']
         : (() => {
-          if (input !== undefined && input['primaryColor'] !== undefined) errors['theme.primaryColor'] = 'must be a valid hex color'
+          if (input !== undefined && input['primaryColor'] !== undefined) errors['theme.primaryColor'] = 'must be a valid rgb, hex, hsl or oklch color'
           return defaultConfig.theme?.primaryColor
         })()
   }